python-parser/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
pydantic>=2.5.0
aiohttp>=3.9.0
tqdm>=4.66.0
